Reroute Schedule Page
