+++ src/pages/help.tsx
@@ -1,0 +1,12 @@
+import Header from "../components/Header";
+
+const Help = () => {
+  return (
+    <>
+      <Header></Header>
+      Help Page
+    </>
+  );
+};
+
+export default Help;
